+++ src/core/highlighters/key_value.rs
@@ -1,8 +1,9 @@
+use super::RegexExt;
 use crate::core::config::KeyValueConfig;
 use crate::core::highlighter::Highlight;
 use memchr::memchr;
 use nu_ansi_term::Style as NuStyle;
-use regex::{Captures, Error, Regex, RegexBuilder};
+use regex::{Error, Regex, RegexBuilder};
 use std::borrow::Cow;
 use std::fmt::Write as _;
 
@@ -31,17 +32,15 @@ impl Highlight for KeyValueHighlighter {
             return Cow::Borrowed(input);
         }
 
-        self.regex.replace_all(input, |captures: &Captures| {
-            let space_or_start = captures.name("space_or_start").map(|s| s.as_str()).unwrap_or_default();
-            let mut buf = String::with_capacity(space_or_start.len() + 32);
+        self.regex.replace_all_cow(input, |caps, buf| {
+            let space_or_start = caps.name("space_or_start").map(|s| s.as_str()).unwrap_or_default();
             buf.push_str(space_or_start);
-            if let Some(k) = captures.name("key") {
+            if let Some(k) = caps.name("key") {
                 let _ = write!(buf, "{}", self.key.paint(k.as_str()));
             }
-            if let Some(e) = captures.name("equals") {
+            if let Some(e) = caps.name("equals") {
                 let _ = write!(buf, "{}", self.separator.paint(e.as_str()));
             }
-            buf
         })
     }
 }